Revert "string-list API: change "nr" and "alloc" to "size_t""
This reverts commit 3a4d800023b612293c1806767a064ce0ca93eccc.

+++ builtin/receive-pack.c
@@ -822,12 +822,12 @@ static int run_and_feed_hook(const char *hook_name, feed_fn feed,
 	proc.trace2_hook_name = hook_name;
 
 	if (feed_state->push_options) {
-		size_t i;
+		int i;
 		for (i = 0; i < feed_state->push_options->nr; i++)
 			strvec_pushf(&proc.env_array,
-				     "GIT_PUSH_OPTION_%"PRIuMAX"=%s", i,
+				     "GIT_PUSH_OPTION_%d=%s", i,
 				     feed_state->push_options->items[i].string);
-		strvec_pushf(&proc.env_array, "GIT_PUSH_OPTION_COUNT=%"PRIuMAX"",
+		strvec_pushf(&proc.env_array, "GIT_PUSH_OPTION_COUNT=%d",
 			     feed_state->push_options->nr);
 	} else
 		strvec_pushf(&proc.env_array, "GIT_PUSH_OPTION_COUNT");
